clean up incident form
