+++ site/gatsby-site/src/components/forms/IncidentForm/index.js
@@ -1,0 +1,3 @@
+import IncidentForm from './Form';
+
+export default IncidentForm;
